pyproject.toml
[project]
name = "remote-db-mcp-server"
version = "0.1.0"
description = "Remote DB MCP Server"

dependencies = [
    "azure-cosmos",
    "azure-identity",
    "python-dotenv",
    "httpx",
    "mcp",
    "langchain-openai",
    "langgraph",
    "langchain-core",
    "langchain-mcp-adapters",
    "debugpy",
    "fastapi",
    "uvicorn[standard]",
]

[project.scripts]
remote-db-mcp-server = "src.__main__:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
    packages = ["src/remote_db_mcp_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]